first_data
I am building a robot to analyze Ethereum data that will analyze the Ethereum market live and up-to-date using data from the last year. This version is under construction and is in its early stages. It is going to become a powerful tool that you can log in with its own dedicated application and analyze the market.